requirements.txt
azure-storage-blob
azure-identity
azure-keyvault-secrets
azure-keyvault-certificates
azure-keyvault-keys
python-dotenv
torchxrayvision
torchcam
joblib
